# src/app/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MeasureX API",
    "description": "Construction takeoff and measurement platform API",
    "version": "1.0.0",
    "contact": {
      "name": "MeasureX Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative API base"
    }
  ],
...
  "paths": {
...
    "/api/ai-takeoff": {
      "post": {
        "tags": [
          "AI Takeoff"
        ],
        "summary": "Standalone AI takeoff (no project context)",
        "operationId": "standaloneAiTakeoff",
        "description": "Analyzes a base64-encoded blueprint image for construction elements\nwithout requiring a project. Uses GPT-4o-mini.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "imageBase64",
                  "pageWidth",
                  "pageHeight"
                ],
                "properties": {
                  "imageBase64": {
                    "type": "string",
                    "description": "Base64-encoded image"
                  },
                  "pageWidth": {
                    "type": "number"
                  },
                  "pageHeight": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detected elements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "classification": {
                            "type": "string"
                          },
                          "points": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "x": {
                                  "type": "number"
                                },
                                "y": {
                                  "type": "number"
                                }
                              }
                            }
                          },
                          "color": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/image-search": {
      "post": {
        "tags": [
          "Search"
        ],
        "summary": "Search for construction images",
        "operationId": "imageSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "query"
                ],
                "properties": {
                  "query": {
                    "type": "string",
                    "example": "concrete foundation"
                  },
                  "projectId": {
                    "type": "string",
                    "description": "Optional project ID to include project sheet results"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "thumbUrl": {
                            "type": "string"
                          },
                          "fullUrl": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string"
                          },
                          "pageNumber": {
                            "type": "integer"
                          },
                          "sheetName": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}